requirements.txt
streamlit==1.36.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
tensorflow==2.16.2
scikit-learn==1.5.0
keras-tuner==1.4.7
google-generativeai==0.7.1
fpdf==1.7.2
requests==2.32.3
ta==0.11.0